Changed Live feed section
Tweaked live feed section so that it picks up better results, has five
section. In four sections the results now come from API and json, so
less worry of html breaking, created a live feed search (but uses html
due to limitations of api and live feeds) and created a live feed
parsing program that may be able to be used for other purposes.